interface - class 네이밍수정
